refactor: share telemetry schema across sdk and bench
