+++ packages/telemetry/CHANGELOG.md
@@ -1,0 +1,6 @@
+# @computesdk/telemetry
+
+## 0.1.0
+
+- Initial release.
+- Add shared telemetry event types and transport helpers.
