+++ packages/telemetry/README.md
@@ -1,0 +1,5 @@
+# @computesdk/telemetry
+
+Shared telemetry schema and transport utilities for ComputeSDK packages.
+
+This package is intentionally minimal: it defines event types and helpers for ID generation, runtime detection, error-code normalization, and event emission.
